+++ src/EA.Iws.Core/Movement/Bulk/BulkMovementContentRules.cs
@@ -4,6 +4,8 @@
 
     public enum BulkMovementContentRules
     {
+        [Display(Name = "Shipment number {0}: there is missing data.")]
+        MissingData,
         [Display(Name = "Shipment number {0}: the actual date of shipment can't be outside of the consent validity period.")]
         ConsentValidity,
         [Display(Name = "Shipment number {0}: is duplicated within the data file.")]
@@ -14,8 +16,6 @@
         HistoricDate,
         [Display(Name = "Shipment number {0}: the date of shipment format must be in dd/mm/yyyy format.")]
         InvalidDateFormat,
-        [Display(Name = "Shipment number {0}: there is missing data.")]
-        MissingData,
         [Display(Name = "You can't create {0} shipment(s) as they are missing a value for shipment number.")]
         MissingShipmentNumbers,
         [Display(Name = "Shipment number {0}: data must only be for notification number {1}")]
