+++ src/EA.Iws.Core/Movement/Bulk/IBulkMovementPrenotificationContentRule.cs
@@ -3,7 +3,6 @@
     using System;
     using System.Collections.Generic;
     using System.Threading.Tasks;
-    using Core.Movement.Bulk;
 
     public interface IBulkMovementPrenotificationContentRule
     {
