+++ src/EA.Iws.RequestHandlers/EA.Iws.RequestHandlers.csproj
@@ -450,7 +450,6 @@
     <Compile Include="NotificationMovements\BulkUpload\PrenotificationContentExcessiveShipmentsRule.cs" />
     <Compile Include="NotificationMovements\BulkUpload\PrenotificationContentInvalidPackagingTypeRule.cs" />
     <Compile Include="NotificationMovements\BulkUpload\PrenotificationContentInvalidShipmentNumberRule.cs" />
-    <Compile Include="NotificationMovements\BulkUpload\PrenotificationContentMissingDataRule.cs" />
     <Compile Include="NotificationMovements\BulkUpload\PrenotificationContentMissingShipmentNumberRule.cs" />
     <Compile Include="NotificationMovements\BulkUpload\PrenotificationContentNotificationNumberRule.cs" />
     <Compile Include="NotificationMovements\BulkUpload\PrenotificationContentOnlyNewShipmentsRule.cs" />
